requirements.txt
apscheduler==3.7.0
Flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
slackclient==2.9.3
slackeventsapi==2.2.1
SQLAlchemy==1.4.23